Added persistence

+++ src/main/java/com/giraff/repository/PersonRepositoryManager.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
+import javax.persistence.EntityNotFoundException;
 import javax.persistence.Persistence;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
@@ -11,14 +12,14 @@ import javax.persistence.TypedQuery;
 import com.giraff.model.Person;
 
 
-public class PersonRepositoryManager { //implements PersonRepository {
+public class PersonRepositoryManager implements PersonRepository {
 	
-	private static final String PERSISTENCE_UNIT_NAME = "Person";
+	private static final String PERSISTENCE_UNIT_NAME = "PersonPU";
 	private static EntityManagerFactory factory = Persistence.createEntityManagerFactory(PERSISTENCE_UNIT_NAME);
 	@PersistenceContext 
 	private EntityManager em = factory.createEntityManager();
 
-//   @Override
+   @Override
   	public Person find(String personId) {
        Long id;
 	   try {
@@ -29,56 +30,71 @@ public class PersonRepositoryManager { //implements PersonRepository {
 	   Person person = em.find(Person.class, id);
        return person;
     }
-//    @Override
+    @Override
     public List<Person> findAll() {
         TypedQuery<Person> query = em.createQuery(
                 "SELECT p FROM Person p ORDER BY p.familyName", Person.class);
             return query.getResultList();
     }
 
-//    @Override
+    @Override
     public Person persist(Person person) {
     	long id;
-        try{
-        	  TypedQuery<Long> query = em.createQuery(
-        		      "SELECT Max(p.id) FROM Person p", Long.class);
-        		  id = query.getSingleResult();
-        		  
-        }catch(Exception e){
-        	id = 0;
-        }
-        try{
-      		  person.setId(id);
-      		  em.persist(person);
-      		  em.flush();
-      }catch(Exception e){
-      	System.out.println("Couldn't persist");
-      }
-        finally{
-        	try {
-				em.close();
-			} catch (Exception e) {
-			}
-        }
-        return person;
+    	try{
+    		TypedQuery<Long> query = em.createQuery(
+    				"SELECT Max(p.id) FROM Person p", Long.class);
+    		id = query.getSingleResult();
+    		id++; //add one to id
+
+    	}catch(Exception e){
+    		//No rows in db, id = 0
+    		id = 0;
+    	}
+		person.setId(id);
+    	try{
+    		em.getTransaction().begin();
+    		em.persist(person);
+//    		em.flush();
+    		em.getTransaction().commit();
+    	}catch(Exception e){
+    		System.out.println("Couldn't persist");
+    		throw new RuntimeException(e);
+    	}
+    	finally{
+    		try {
+    			em.close();
+    		} catch (Exception e) {
+        		System.out.println("Couldn't close connection");
+        		throw new RuntimeException(e);
+    		}
+    	}
+    	return person;
     }
 
-//    @Override
+    @Override
     public Person merge(Person person) {
         try{
+    		em.getTransaction().begin();
         	em.merge(person);
-        	em.flush();
+    		em.getTransaction().commit();
+        }catch(EntityNotFoundException e){
+        	em.persist(person);
+    		em.getTransaction().commit();
         }catch(Exception e){
+    		System.out.println("Couldn't merge");
+    		throw new RuntimeException(e);
         }
         finally{
         	try {
 				em.close();
 			} catch (Exception e) {
+        		System.out.println("Couldn't close connection");
+        		throw new RuntimeException(e);
 			}
         }
     	return person;
     }
-//    @Override
+    @Override
     public void delete(String personId) {
     	Long id;
     	try {
@@ -86,17 +102,24 @@ public class PersonRepositoryManager { //implements PersonRepository {
     	} catch (NumberFormatException e) {
     		throw e;
     	}
-
     	try{
+    		em.getTransaction().begin();
     		Person person = em.find(Person.class, id);
-    		em.remove(person);
-    		em.flush();
+    		if (person != null) {
+        		em.remove(person);
+    		}
+//    		em.flush();
+    		em.getTransaction().commit();
     	}catch(Exception e){
+    		System.out.println("Couldn't delete");
+    		throw new RuntimeException(e);
     	}
     	finally{
     		try {
     			em.close();
     		} catch (Exception e) {
+        		System.out.println("Couldn't close connection");
+        		throw new RuntimeException(e);
     		}
     	}
     	return;
